% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_label.R
\name{get_survey_label}
\alias{get_survey_label}
\title{Get survey label from survey name}
\usage{
get_survey_label(survey, col)
}
\arguments{
\item{survey}{A survey sheet from Kobo (already split with columns label and name present).}

\item{col}{A column.}
}
\value{
The label or NA if it does not exist.
}
\description{
Get survey label from survey name
}
