% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_labels.R
\name{get_survey_labels}
\alias{get_survey_labels}
\title{Get survey labels from survey names}
\usage{
get_survey_labels(survey, ..., output_df = TRUE)
}
\arguments{
\item{survey}{A survey sheet from Kobo (already split with columns label and name present).}

\item{...}{Column names, quoted or unquoted.}

\item{output_df}{Boolean. Should the output be a data frame or a vector? Default to TRUE.}
}
\value{
The labels as a vectir.
}
\description{
Get survey labels from survey names
}
