% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_type.R
\name{get_survey_type}
\alias{get_survey_type}
\alias{get_survey_select_one}
\alias{get_survey_select_multiple}
\alias{get_survey_calculate}
\alias{get_survey_integer}
\title{Get questions by type from the survey sheet}
\usage{
get_survey_type(survey, question_type)

get_survey_select_one(survey)

get_survey_select_multiple(survey)

get_survey_calculate(survey)

get_survey_integer(survey)
}
\arguments{
\item{survey}{The survey sheet from Kobo (with column "type" split). See \code{split_survey()}}

\item{question_type}{Usually "integer", "select_one" or "select_multiple".}
}
\value{
A character vector of select_one questions.
}
\description{
\code{get_survey_type()} is a generic; the other functions get the specified obvious type from the survey sheet.
}
\details{
\code{survey} should have a split type column with types of variables such as "select_one", "select_multiple", etc.
}
