% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_all.R
\name{label_all}
\alias{label_all}
\title{Label all select_one and select_multiple questions}
\usage{
label_all(df, survey, choices, id_col)
}
\arguments{
\item{df}{Some kobo data.}

\item{survey}{Some survey sheet, with a split 'type' column, e.g. with \code{split_survey(survey, type)}. It must have columns 'list_name', 'type', 'label', and 'name'.}

\item{choices}{The corresponding choices sheet; it must have columns 'name' and 'label'.}

\item{id_col}{The survey id column; usually uuid.}
}
\value{
A labeled data frame.
}
\description{
It uses \code{label_all_select_one()} and \code{label_all_select_multiple()}in the background.
}
