% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_columns.R
\name{label_columns}
\alias{label_columns}
\title{Label Kobo data column names as attributes}
\usage{
label_columns(df, survey, name_as_label = TRUE)
}
\arguments{
\item{df}{Some Kobo data.}

\item{survey}{Some survey sheet, with a split 'type' column, e.g. with \code{split_survey(type)}. It must have columns 'list_name', 'type', 'label', and 'name'.}

\item{name_as_label}{Default to TRUE. Should the variable name be used as the label if label is missing?}
}
\value{
Labeled columns as attributes.
}
\description{
Label Kobo data column names as attributes
}
