% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_select_multiple.R
\name{label_select_multiple}
\alias{label_select_multiple}
\alias{label_all_select_multiple}
\title{Label a select_multiple question}
\usage{
label_select_multiple(df, survey, choices, id_col, col, return_df = NULL)

label_all_select_multiple(df, survey, choices, id_col)
}
\arguments{
\item{df}{Some kobo data.}

\item{survey}{Some survey sheet, with a split 'type' column, e.g. with \code{split_survey(type)} and at least one 'label' column.}

\item{choices}{The corresponding choices sheet; it must have columns name and label.}

\item{id_col}{The id column; usually uuid.}

\item{col}{The select_one column to label.}

\item{return_df}{NULL returns the updated data frame; "id_col" returns a data frame with only col and id_col; and, "vector" returns.}
}
\value{
A labeled data frame, sub-data frame or vector for \code{label_select_multiple()}; a labeled data frame for \code{label_all_select_multiple()}.
}
\description{
\code{label_select_multiple()} labels one variable. \code{label_all_select_multiple()} label all select_multiple variables that exists in the survey sheet and in the data set.
}
